third_party/WebKit/LayoutTests/fast/events/touch/compositor-touch-hit-rects-global-expected.txt

This tests verifies the hit test regions given to the compositor in the simple case where touch handles cover (or nearly cover) the entire document. It can only be run in DumpRenderTree.

document: #document (0, 0, 800, 600)

html: #document (0, 0, 800, 600)

body: #document (0, 0, 800, 600)

webPageOverlay: no rects
